Reorganize project
- All code is not in the rtp (temporary name) package python.
- Add readme to the example directory
- Update main readme
- Remove readme placeholders in server and client directories
- Remove non-working build system, lets add it later if needed
- Fix setup.py so we can create source tarball - this is good
  enough for delivery.
