--- MANIFEST
@@ -1,19 +1,0 @@
-# file GENERATED by distutils, do NOT edit
-AUTHORS
-COPYING
-README
-setup.py
-start_client
-start_server
-client/__init__.py
-client/config.py
-client/game.py
-client/main.py
-common/message.py
-components/__init__.py
-components/car.py
-components/component.py
-components/track.py
-server/game.py
-server/gameServer.py
-server/main.py
