+++ examples/your_driver.py
@@ -1,4 +1,4 @@
-from common import obstacles, actions
+from rtp.common import obstacles, actions
 
 
 def drive(your_car_location, check_for_obstacle_func):
