+++ rtp-client
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+from rtp.client import main
+main.main()
