+++ rtp-server
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+from rtp.server import main
+main.main()
