+++ rtp/client/game.py
@@ -2,7 +2,7 @@ from collections import namedtuple
 from twisted.internet import reactor
 from twisted.internet import task
 import pygame
-from common import config, message
+from rtp.common import config, message
 import track, car, component
 
 author = 'gickowic'
