+++ rtp/common/error.py
