+++ rtp/common/obstacles.py
