+++ rtp/res/obstacles/obstacle3.png
