+++ rtp/res/obstacles/obstacle4.png
