+++ rtp/res/obstacles/obstacle6.png
