+++ setup.py
@@ -1,16 +1,12 @@
 from distutils.core import setup
 
 setup(name='RaananaTiraProject',
-      version='1.0',
+      version='0.1',
       license="GNU GPLv2+",
       description="game",
-      py_modules=['server/game', 'server/gameServer', 'server/main',
-                  'client/game', 'client/config', 'client/__init__',
-                  'client/main',
-                  'components/car', 'components/component',
-                  'components/__init__', 'components/track',
-                  'common/message'],
+      packages=['rtp', 'rtp.server', 'rtp.client', 'rtp.common'],
+      package_data={'rtp': ['res/*/*.png']},
       author='Yaniv Bronhaim',
       author_email='ybronhei@redhat.com',
       url="https://github.com/emesika/RaananaTiraProject",
-      scripts=["start_client", "start_server"])
+      scripts=["rtp-client", "rtp-server"],)
