--- start_client
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-PYTHONPATH=. python client/main.py $@
