--- start_server
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-PYTHONPATH=. python server/main.py
